backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.0
python-dotenv==1.0.1
aiohttp==3.11.10

# AI providers (pluggable)
anthropic==0.40.0
openai==1.57.2
ollama==0.4.4

# IKEA Dirigera integration
dirigera==0.9.3

# Matter / Thread (python-matter-server client)
python-matter-server[client]==6.6.0

# Utilities
structlog==24.4.0
python-multipart==0.0.19
